renamed vector components
