+++ sites/geohub/src/components/controls/VectorHeatmap.svelte
