+++ sites/geohub/src/components/controls/VectorPolygon.svelte
